gh-75. Move auth actions to authHook
