+++ apps/web-portal-client/lib/features/auth/containers/LoginPage.jsx
@@ -1,36 +1,21 @@
-import React, {useContext, useEffect} from 'react'
-import {useSelector, useDispatch} from 'react-redux'
+import React, {useEffect} from 'react'
+import {useSelector} from 'react-redux'
 import {useHistory} from 'react-router-dom'
 import Spinner from '../../../components/Spinner'
 import {selectLoginState, selectAuthError, selectAuthToken} from '../selectors'
-import {AppContext} from '../../../context'
-import {error, logging, login} from '../slices'
+import {useAuth} from '../hooks'
 
 export function LoginPage() {
-  const dispatch = useDispatch()
   const history = useHistory()
   const isLogging = useSelector(selectLoginState)
   const loginError = useSelector(selectAuthError)
   const token = useSelector(selectAuthToken)
-  const {api} = useContext(AppContext)
+  const auth = useAuth()
   useEffect(() => {
     if (!isLogging && !loginError && token) {
       history.push('/home')
     }
   })
-
-  function handleSubmit(e) {
-    e.preventDefault()
-    dispatch(logging())
-    api.auth
-      .login()
-      .then((result) => {
-        dispatch(login(result))
-      })
-      .catch((err) => {
-        dispatch(error(err))
-      })
-  }
 
   return (
     <div className="min-h-screen flex items-center justify-center bg-gray-50 py-12 px-4 sm:px-6 lg:px-8">
@@ -96,14 +81,15 @@ export function LoginPage() {
               className="group relative w-full flex justify-center py-2 px-4 border border-transparent text-sm font-medium rounded-md text-white bg-indigo-600 hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500"
               onClick={(event) => {
                 // todo: pass email and passwoord to action as parameter dispatch(userLogin(params))
-                handleSubmit(event)
+                event.preventDefault()
+                auth.login()
               }}
             >
               <span className="absolute left-0 inset-y-0 flex items-center pl-3" />
               <Spinner show={isLogging} />
               Sign in
             </button>
-            {error && <div>Invalid password or email</div>}
+            {loginError && <div>Invalid password or email</div>}
           </div>
         </form>
       </div>
